Address PR #290 review: split HTTP types per-file, self-contain AbortSignal
- Split IHttpClient.ts into one-type-per-file (HttpResponseType,
  IHttpRequestConfig, IHttpResponseEnvelope, IHttpClient) per AGENTS.md
  one-type-per-file convention; update http/ and root barrels.
- Replace the global AbortSignal reference on IHttpRequestConfig.signal
  with a local structural IAbortSignalLike so the published contract
  stays self-contained for consumers lacking dom/@types/node libs; a real
  AbortSignal remains assignable.
