+++ packages/types/src/http/HttpResponseType.ts
@@ -1,0 +1,8 @@
+/**
+ * DRAFT (tronrelic#289) — supporting type for the proposed outbound HTTP client
+ * contract. See IHttpClient.ts for full status and caveats. Not yet wired into
+ * IPluginContext.http.
+ */
+
+/** Response body interpretation requested from the transport. */
+export type HttpResponseType = 'json' | 'text' | 'stream' | 'arraybuffer';
