+++ packages/types/src/http/index.ts
@@ -13,4 +13,8 @@ export type { IHttpResponse } from './IHttpResponse.js';
 export type { IHttpNext } from './IHttpNext.js';
 // DRAFT (tronrelic#289) — proposed outbound HTTP client contract, not yet wired
 // into IPluginContext.http. See IHttpClient.ts for status and caveats.
-export type { IHttpClient, IHttpRequestConfig, IHttpResponseEnvelope, HttpResponseType } from './IHttpClient.js';
+export type { IHttpClient } from './IHttpClient.js';
+export type { IHttpRequestConfig } from './IHttpRequestConfig.js';
+export type { IHttpResponseEnvelope } from './IHttpResponseEnvelope.js';
+export type { HttpResponseType } from './HttpResponseType.js';
+export type { IAbortSignalLike } from './IAbortSignalLike.js';
