+++ packages/types/src/index.ts
@@ -8,7 +8,7 @@ export type { IHttpRequest, IHttpResponse, IHttpNext } from './http/index.js';
 // DRAFT (tronrelic#289) — proposed core-owned HTTP client contract; additive and
 // unwired. Will replace the leaked axios AxiosInstance on IPluginContext.http
 // after the consumer review tracked in the issue.
-export type { IHttpClient, IHttpRequestConfig, IHttpResponseEnvelope, HttpResponseType } from './http/index.js';
+export type { IHttpClient, IHttpRequestConfig, IHttpResponseEnvelope, HttpResponseType, IAbortSignalLike } from './http/index.js';
 export type { IAuthSession, IAuthSessionUser, IHasAuthSession } from './auth/index.js';
 export { ADMIN_GROUP_ID, isLoggedIn, isAnonymous, isInGroup, isAdmin, hasPrimaryWallet } from './auth/index.js';
 // ILogger removed - use ISystemLogService instead (exported from './system-log/index.js')
